pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "versioningit"]
build-backend = "setuptools.build_meta"


# ----------------------------------------- setuptools -------------------------------------------
[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true


# ----------------------------------------- Project Metadata -------------------------------------
#
[project]
version = "0.0.0.dev4"
name = "PySerials"
dependencies = [
    "jsonschema >= 4.21.0, < 5",
    "ruamel.yaml >= 0.17.32, < 0.18", # https://yaml.readthedocs.io/en/stable/
    "ruamel.yaml.string >= 0.1.1, < 1",
    "tomlkit >= 0.11.8, < 0.12", # https://tomlkit.readthedocs.io/en/stable/,
    "markitup",
    "ansi-sgr",
    "ExceptionMan"
]
requires-python = ">=3.10"
